+++ R/automaticCOTANObjectCreation.R
@@ -22,7 +22,7 @@ NULL
 #' @param calcCoex a Boolean to determine whether to calculate the genes' `COEX`
 #'   or stop just before at the [estimateDispersionBisection()] step
 #' @param optimizeForSpeed Boolean; when `TRUE` `COTAN` tries to use the `torch`
-#'   library to run the matrix calcualtions. Otherwise, or when the library is
+#'   library to run the matrix calculations. Otherwise, or when the library is
 #'   not available will run the slower legacy code
 #' @param deviceStr On the `torch` library enforces which device to use to run
 #'   the calculations. Possible values are `"cpu"` to us the system *CPU*,
@@ -198,7 +198,7 @@ setMethod(
 #' @param calcCoex a Boolean to determine whether to calculate the genes' `COEX`
 #'   or stop just before at the [estimateDispersionBisection()] step
 #' @param optimizeForSpeed Boolean; when `TRUE` `COTAN` tries to use the `torch`
-#'   library to run the matrix calcualtions. Otherwise, or when the library is
+#'   library to run the matrix calculations. Otherwise, or when the library is
 #'   not available will run the slower legacy code
 #' @param deviceStr On the `torch` library enforces which device to use to run
 #'   the calculations. Possible values are `"cpu"` to us the system *CPU*,