+++ R/calculateCoex-method.R
@@ -1100,7 +1100,7 @@ calculateCoex_Torch <- function(objCOTAN, deviceStr = "cpu") {
 #' @param actOnCells Boolean; when `TRUE` the function works for the cells,
 #'   otherwise for the genes
 #' @param optimizeForSpeed Boolean; when `TRUE` `COTAN` tries to use the `torch`
-#'   library to run the matrix calcualtions. Otherwise, or when the library is
+#'   library to run the matrix calculations. Otherwise, or when the library is
 #'   not available will run the slower legacy code
 #' @param deviceStr On the `torch` library enforces which device to use to run
 #'   the calculations. Possible values are `"cpu"` to us the system *CPU*,