+++ R/data.R
@@ -72,3 +72,41 @@ NULL
 #' @rdname Datasets
 #'
 "test.dataset.clusters2"
+
+
+
+#' @title Installing torch R library (on Linux)
+#'
+#' @description A brief explanation of how to install the torch package on
+#'   `WSL2` (Windows Subsystem for Linux), but it might work the same for other
+#'   `Linux` systems. Naturally it makes a difference whether one wants to
+#'   install support only for the `CPU` or also have the system `GPU` at the
+#'   ready!
+#'
+#' @description The main resources to install `torch` is
+#'   \url{https://torch.mlverse.org/docs/articles/installation.html} or
+#'   \url{https://cran.r-project.org/web/packages/torch/vignettes/installation.html}
+#'
+#' @details For the `CPU`-only support one need to ensure that also numeric
+#'   libraries are installed, like `BLAS` and `LAPACK` and/or `MKL` if your
+#'   `CPU` is from *Intel*. Otherwise `torch` will be stuck at using a single
+#'   core for all computations.
+#'
+#' @details For the `GPU`, currently only `cuda` devices are supported. Moreover
+#'   only some specific versions of `cuda` (and corresponding `cudnn`) are
+#'   effectively usable, so one needs to install them to actually use the `GPU`.
+#'
+#'   As of today only `cuda` 11.7 and 11.8 are supported, but check the `torch`
+#'   documentation for more up-to-date information. Before downgrading your
+#'   `cuda` version, please be aware that it is possible to maintain separate
+#'   main versions of `cuda` at the same time on the system: that is one can
+#'   have installed both 11.8 and a 12.4 `cuda` versions on the same system.
+#'
+#'   Below a link to install `cuda` 11.8 for `WSL2` given: use a local installer
+#'   to be sure the wanted `cuda` version is being installed, and not the latest
+#'   one: [cuda 11.8 for
+#'   WSL2](https://developer.nvidia.com/cuda-11-8-0-download-archive?target_os=Linux&target_arch=x86_64&Distribution=WSL-Ubuntu&target_version=2.0&target_type=deb_local)
+#'
+#' @name Installing_torch
+#'
+NULL