+++ R/findClustersMarkers.R
@@ -18,7 +18,6 @@
 #'   of the *clusters* of the *clusterization*
 #' @param method *p-value* multi-test adjustment method. Defaults to
 #'   `"bonferroni"`; use `"none"` for no adjustment
-#' @param cores number of cores to use. Default is 1.
 #'
 #' @returns `findClustersMarkers()` returns a `data.frame` containing `n` genes
 #'   for each *cluster* scoring top/bottom `COEX` scores. The `data.frame` also
@@ -41,14 +40,14 @@
 #'
 #' @examples
 #' clMarkers <- findClustersMarkers(objCOTAN, markers = list(),
-#'                                  clusters = clusters, cores = 6L)
+#'                                  clusters = clusters)
 #'
 #' @rdname HandlingClusterizations
 #'
 findClustersMarkers <- function(
     objCOTAN, n = 10L, markers = NULL,
     clName = "", clusters = NULL,
-    coexDF = NULL, method = "bonferroni", cores = 1L) {
+    coexDF = NULL, method = "bonferroni") {
   logThis("findClustersMarkers - START", logLevel = 2L)
 
   marks <- unlist(markers)
@@ -66,7 +65,7 @@ findClustersMarkers <- function(
       coexDF <- getClusterizationData(objCOTAN, clName = clName)[["coex"]]
     }
     if (is_empty(coexDF)) {
-      coexDF <- DEAOnClusters(objCOTAN, clusters = clusters, cores = cores)
+      coexDF <- DEAOnClusters(objCOTAN, clusters = clusters)
     }
   }