+++ tests/testthat/test-cellsUniformClustering.R
@@ -12,7 +12,7 @@ test_that("Cell Uniform Clustering", {
                                sampleCondition = "test")
 
   obj <- proceedToCoex(obj, calcCoex = FALSE,
-                       cores = 12L, saveObj = TRUE, outDir = tm)
+                       cores = 6L, saveObj = TRUE, outDir = tm)
 
   GDIThreshold <- 1.46
   initialResolution <- 0.8
@@ -20,8 +20,8 @@ test_that("Cell Uniform Clustering", {
     c(clusters, coexDF) %<-%
       cellsUniformClustering(obj, GDIThreshold = GDIThreshold,
                              initialResolution = initialResolution,
-                             cores = 12L, saveObj = TRUE,
-                             outDir = tm)
+                             cores = 6L, optimizeForSpeed = TRUE,
+                             deviceStr = "cuda", saveObj = TRUE, outDir = tm)
   })
 
   expect_true(file.exists(file.path(tm, "test", "reclustering",
@@ -42,14 +42,16 @@ test_that("Cell Uniform Clustering", {
   expect_identical(reorderClusterization(obj)[["clusters"]], clusters)
 
   firstCl <- clusters[[1L]]
-  c(isUniform, fracAbove, lastPerc) %<-%
+  c(isUniform, fracAbove, firstPerc, clSize) %<-%
     checkClusterUniformity(obj, GDIThreshold = GDIThreshold,
-                           cluster = paste0("Cluster_", firstCl),
+                           clusterName = paste0("Cluster_", firstCl),
                            cells = names(clusters)[clusters == firstCl],
+                           optimizeForSpeed = TRUE, deviceStr = "cpu",
                            saveObj = TRUE, outDir = tm)
   expect_true(isUniform)
   expect_lte(fracAbove, 0.01)
-  expect_lte(lastPerc, GDIThreshold)
+  expect_lte(firstPerc, GDIThreshold)
+  expect_identical(clSize, sum(clusters == firstCl))
 
   clusters2 <- factor(clusters, levels = c("-1", levels(clusters)))
   clusters2[1L:50L] <- "-1"
@@ -83,7 +85,8 @@ test_that("Cell Uniform Clustering", {
     splitList <- cellsUniformClustering(obj, GDIThreshold = GDIThreshold,
                                         initialResolution = initialResolution,
                                         initialClusters = exactClusters,
-                                        cores = 12L, saveObj = TRUE,
+                                        cores = 6L, optimizeForSpeed = FALSE,
+                                        deviceStr = "cpu", saveObj = TRUE,
                                         outDir = tm)
   })
 
@@ -108,8 +111,7 @@ test_that("Cell Uniform Clustering", {
   primaryMarkers <-
     c("g-000010", "g-000020", "g-000030", "g-000300", "g-000330",
       "g-000510", "g-000530", "g-000550", "g-000570", "g-000590")
-  clMarkersDF2 <- findClustersMarkers(obj, markers = primaryMarkers,
-                                      cores = 12L)
+  clMarkersDF2 <- findClustersMarkers(obj, markers = primaryMarkers)
 
   expect_identical(colnames(clMarkersDF2), colnames(clMarkersDF))
   expect_identical(clMarkersDF2[, -6L], clMarkersDF[, -6L])
@@ -129,7 +131,7 @@ test_that("Cell Uniform Clustering", {
     temp.obj <- dropGenesCells(objCOTAN = obj,
                                cells = getCells(obj)[cellsToDrop])
 
-    temp.obj <- proceedToCoex(temp.obj, cores = 12L, saveObj = FALSE)
+    temp.obj <- proceedToCoex(temp.obj, cores = 6L, saveObj = FALSE)
     gc()
 
     GDI_data <- calculateGDI(temp.obj)