+++ NEWS.md
@@ -1,3 +1,9 @@
+## 2.5.0
+
+Added support for the `torch` library to help with the heavy lifting
+calculations of the *genes' COEX* matrix, with consequent substantial speed-up,
+especially when a **GPU** is available on the system
+
 ## COTAN 2.3.6
 
 Refactored `DEAOnCluster()` to make it run faster.